Revert search query for owner and member
It's used to be full text search for all searchable files which includes
owner and members, but now we used a more sophisticated analyzer with
some chess synonyms filter, which doesn't make sense to search with
owner and/or members. So I removed them from full text search. But it
seems people rely on that for manage their studies, most probably to
working with students.